Moved env config into environments folder
