+++ config/__init__.py
@@ -7,11 +7,15 @@ if not FLASK_ENV:
 
 match FLASK_ENV:
     case "development":
-        from config.development import DevelopmentConfig as Config  # noqa
+        from config.environments.development import (  # noqa
+            DevelopmentConfig as Config,  # noqa
+        )  # noqa
     case "unit_test":
-        from config.unit_test import UnitTestConfig as Config  # noqa
+        from config.environments.unit_test import (  # noqa
+            UnitTestConfig as Config,  # noqa
+        )  # noqa
     case _:
-        from config.default import DefaultConfig as Config  # noqa
+        from config.environments.default import DefaultConfig as Config  # noqa
 
 try:
     Config.pretty_print()
