+++ config/environments/development.py
@@ -1,4 +1,4 @@
-from config.default import DefaultConfig
+from config.environments.default import DefaultConfig
 from fsd_utils import configclass
